Make PropertyRule public and move to platform
